+++ player/src/main/java/com/tpstream/player/ui/TPStreamPlayerView.kt
@@ -8,8 +8,9 @@ import android.content.DialogInterface
 import android.content.res.ColorStateList
 import android.graphics.Color
 import android.util.AttributeSet
-import android.util.Log
-import android.view.*
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import android.widget.*
 import androidx.annotation.ColorInt
 import androidx.core.view.isVisible
@@ -20,7 +21,6 @@ import androidx.lifecycle.ViewModelStore
 import androidx.lifecycle.ViewModelStoreOwner
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import androidx.viewpager2.widget.ViewPager2.ScrollState
 import com.tpstream.player.*
 import com.tpstream.player.constants.PlaybackSpeed
 import com.tpstream.player.data.Asset
@@ -29,6 +29,7 @@ import com.tpstream.player.data.source.local.DownloadStatus
 import com.tpstream.player.databinding.TpStreamPlayerViewBinding
 import com.tpstream.player.offline.TpStreamDownloadManager
 import com.tpstream.player.ui.AdvancedResolutionSelectionSheet.OnAdvanceResolutionClickListener
+import com.tpstream.player.ui.adapter.PlaybackSpeedAdapter
 import com.tpstream.player.ui.viewmodel.VideoViewModel
 import com.tpstream.player.util.MarkerState
 import com.tpstream.player.util.NetworkClient.Companion.makeHeadRequest
@@ -63,7 +64,7 @@ class TPStreamPlayerView @JvmOverloads constructor(
     private var noticeScreenLayout: LinearLayout? = null
     private var noticeMessage: TextView? = null
     private val coroutineScope = CoroutineScope(Dispatchers.IO)
-
+    var popupWindow: PopupWindow? = null
 
     init {
         registerDownloadListener()
@@ -107,8 +108,6 @@ class TPStreamPlayerView @JvmOverloads constructor(
         addView(noticeScreenLayout)
     }
 
-    private var popupMenu: PopupMenu? = null
-
     private fun initializeControllerVisibilityListener(){
         playerView.setControllerVisibilityListener(androidx.media3.ui.PlayerView.ControllerVisibilityListener {
             if(!playerView.isControllerFullyVisible){
@@ -124,8 +123,6 @@ class TPStreamPlayerView @JvmOverloads constructor(
         }
     }
 
-    var popupWindow: PopupWindow? = null
-
     private fun showPlaybackSpeedPopupWindow(anchor: View) {
         val inflater = anchor.context.getSystemService(LAYOUT_INFLATER_SERVICE) as LayoutInflater
         val popupView = inflater.inflate(R.layout.popup_window_layout, null)
@@ -138,7 +135,17 @@ class TPStreamPlayerView @JvmOverloads constructor(
         val recyclerView: RecyclerView = popupView.findViewById(R.id.recycler_view)
         recyclerView.layoutManager = LinearLayoutManager(anchor.context)
 
-        val adapter = CustomAdapter(anchor.context, PlaybackSpeed.values().toList())
+        val adapter = PlaybackSpeedAdapter(
+            anchor.context,
+            player.exoPlayer.playbackParameters.speed,
+            PlaybackSpeed.values().toList()
+        ) { playbackSpeed ->
+            player.exoPlayer.playbackParameters =
+                player.exoPlayer.playbackParameters.withSpeed(playbackSpeed.value)
+            setPlaybackSpeedText(playbackSpeed.value)
+            popupWindow?.dismiss()
+        }
+
         recyclerView.adapter = adapter
 
         popupWindow?.height = (playerView.height * 0.75).toInt()
@@ -581,43 +588,6 @@ class TPStreamPlayerView @JvmOverloads constructor(
             it.start()
         }
     }
-
-    inner class CustomAdapter(
-        private val context: Context,
-        private val items: List<PlaybackSpeed>
-    ) : RecyclerView.Adapter<CustomAdapter.ViewHolder>() {
-
-        inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-            val check: ImageView = itemView.findViewById(R.id.check)
-            val text: TextView = itemView.findViewById(R.id.text)
-        }
-
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-            val view = LayoutInflater.from(context).inflate(R.layout.playback_speed_list_item, parent, false)
-            return ViewHolder(view)
-        }
-
-        override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-            val playbackSpeed = items[position]
-            holder.text.text = if (playbackSpeed.value == 1.0f) "Normal" else playbackSpeed.text
-            if (player.exoPlayer.playbackParameters.speed == playbackSpeed.value) {
-                holder.itemView.isSelected = true;
-                holder.check.visibility = VISIBLE;
-            } else {
-                holder.itemView.isSelected = false;
-                holder.check.visibility = INVISIBLE;
-            }
-            holder.itemView.setOnClickListener {
-                player.exoPlayer.playbackParameters =
-                    player.exoPlayer.playbackParameters.withSpeed(playbackSpeed.value)
-                setPlaybackSpeedText(playbackSpeed.value)
-                popupWindow?.dismiss()
-            }
-        }
-
-        override fun getItemCount(): Int = items.size
-    }
-
 }
 
 internal interface TPStreamPlayerViewCallBack {
