+++ player/src/main/player-mapping-files/kotlin/Media3PlayerMapping.kt
@@ -7,7 +7,6 @@ internal typealias TimeBar = androidx.media3.ui.TimeBar
 internal typealias OnScrubListener = androidx.media3.ui.TimeBar.OnScrubListener
 internal typealias FullscreenButtonClickListener =  androidx.media3.ui.PlayerView.FullscreenButtonClickListener
 internal typealias ExoplayerResourceID = androidx.media3.ui.R.id
-internal typealias ExoplayerLayoutID = androidx.media3.ui.R.layout
 
 // androidx.media3.exoplayer.trackselection
 internal typealias DefaultTrackSelector = androidx.media3.exoplayer.trackselection.DefaultTrackSelector
